refactor(webview-accounts): extract prewarm MRU + phase-hint into testable units (#1233)
Three small extractions that make the prewarm flow + loading UX cover
≥80% on diff:

- `app/src/utils/webviewAccountMru.ts` — `MRU_ACCOUNT_KEY`,
  `PREWARM_MAX_ACCOUNTS`, `readMruAccountId`, `writeMruAccountId`. Pure
  helpers extracted from `Accounts.tsx`. Both read + write swallow
  storage errors so private-mode / sandboxed environments don't break
  the page.
- `app/src/hooks/usePrewarmMostRecentAccount.ts` — single-fire on-mount
  hook that wraps the MRU prewarm gating logic (no MRU id, empty
  accounts, > PREWARM_MAX_ACCOUNTS, MRU is active, MRU already in
  pending/loading/open). `Accounts.tsx` shrinks to one call site.
- `WebviewHost.tsx` — phase-hint timer moved into a child
  `LoadingPhaseHint` component that mounts only while loading. Its
  elapsed counter resets purely via mount/unmount, so the parent
  effect no longer needs the synchronous `setElapsedMs(0)` calls
  CodeRabbit flagged as `react-hooks/set-state-in-effect`. The child
  uses a counter-based interval (no `Date.now()`), so fake-timer tests
  remain deterministic.

No behaviour change for users — every visible UX path (placeholder,
spinner from frame 1, 5s/10s phase hints, timeout overlay, MRU
prewarm gating, MRU writes on rail click + new-account pick) renders
identically.

Refs #1233. Addresses CodeRabbit Quick Win review on PR #1285 +
unblocks Coverage Gate via the test files added in the next commit.
